Added a dialogue for setting the Civ5 game location.
On start a check is now made to see if the current Civ5 config game location is valid. If it isn't a popup window will open and tell the user so and allows the user to open a file browser to locate their Launcher.exe file to set the Civ5 game location. Also added a button in the GUI to allow doing this directly if the user wants to change the game location.

Fixes #2

Signed-off-by: MaloW <lowegrenmarcus@gmail.com>